+++ R/collectors.R
@@ -19,6 +19,7 @@ collector_find <- function(name) {
 
 #' Parse a character vector.
 #'
+#' @family parsers
 #' @param x Character vector of elements to parse.
 #' @param collector Column specification.
 #' @inheritParams read_delim
@@ -29,12 +30,13 @@ collector_find <- function(name) {
 #' x <- c("1", "2", "3", "NA")
 #' parse_vector(x, col_integer())
 #' parse_vector(x, col_double())
-parse_vector <- function(x, collector, na = c("", "NA"), locale = default_locale()) {
+parse_vector <- function(x, collector, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  stopifnot(is.character(x))
   if (is.character(collector)) {
     collector <- collector_find(collector)
   }
 
-  warn_problems(parse_vector_(x, collector, na = na, locale_ = locale))
+  warn_problems(parse_vector_(x, collector, na = na, locale_ = locale, trim_ws = trim_ws))
 }
 
 #' Parse logicals, integers, and reals
@@ -67,26 +69,26 @@ NULL
 
 #' @rdname parse_atomic
 #' @export
-parse_logical <- function(x, na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_logical(), na = na, locale = locale)
+parse_logical <- function(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_logical(), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_atomic
 #' @export
-parse_integer <- function(x, na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_integer(), na = na, locale = locale)
+parse_integer <- function(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_integer(), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_atomic
 #' @export
-parse_double <- function(x, na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_double(), na = na, locale = locale)
+parse_double <- function(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_double(), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_atomic
 #' @export
-parse_character <- function(x, na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_character(), na = na, locale = locale)
+parse_character <- function(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_character(), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_atomic
@@ -137,8 +139,8 @@ col_skip <- function() {
 #' @examples
 #' parse_number("$1000")
 #' parse_number("1,234,567.78")
-parse_number <- function(x, na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_number(), na = na, locale = locale)
+parse_number <- function(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_number(), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_number
@@ -176,8 +178,8 @@ col_number <- function() {
 #' # ISO 8601 date times
 #' guess_parser(c("2010-10-10"))
 #' parse_guess(c("2010-10-10"))
-parse_guess <- function(x, na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, guess_parser(x, locale), na = na, locale = locale)
+parse_guess <- function(x, na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, guess_parser(x, locale), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_guess
@@ -223,8 +225,8 @@ guess_parser <- function(x, locale = default_locale()) {
 #' # Using an argument of `NULL` will generate levels based on values of `x`
 #' x2 <- parse_factor(x, levels = NULL)
 parse_factor <- function(x, levels, ordered = FALSE, na = c("", "NA"),
-                         locale = default_locale(), include_na = TRUE) {
-  parse_vector(x, col_factor(levels, ordered, include_na), na = na, locale = locale)
+                         locale = default_locale(), include_na = TRUE, trim_ws = TRUE) {
+  parse_vector(x, col_factor(levels, ordered, include_na), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_factor
@@ -256,7 +258,8 @@ col_factor <- function(levels, ordered = FALSE, include_na = FALSE) {
 #'     70-99 -> 1970-1999.
 #'   \item Month: "\%m" (2 digits), "\%b" (abbreviated name in current
 #'     locale), "\%B" (full name in current locale).
-#'   \item Day: "\%d" (2 digits), "\%e" (optional leading space)
+#'   \item Day: "\%d" (2 digits), "\%e" (optional leading space),
+#'     "%a" (abbreviated name in current locale).
 #'   \item Hour: "\%H" or "\%I", use I (and not H) with AM/PM.
 #'   \item Minutes: "\%M"
 #'   \item Seconds: "\%S" (integer seconds), "\%OS" (partial seconds)
@@ -300,7 +303,7 @@ col_factor <- function(levels, ordered = FALSE, include_na = FALSE) {
 #' @inheritParams tokenizer_delim
 #' @return A [POSIXct()] vector with `tzone` attribute set to
 #'   `tz`. Elements that could not be parsed (or did not generate valid
-#'   dates) will bes set to `NA`, and a warning message will inform
+#'   dates) will be set to `NA`, and a warning message will inform
 #'   you of the total number of failures.
 #' @family parsers
 #' @export
@@ -363,20 +366,20 @@ col_factor <- function(levels, ordered = FALSE, include_na = FALSE) {
 #' parse_datetime("1979-10-14T1010Z", locale = us_central)
 #' # Your current time zone
 #' parse_datetime("1979-10-14T1010", locale = locale(tz = ""))
-parse_datetime <- function(x, format = "", na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_datetime(format), na = na, locale = locale)
+parse_datetime <- function(x, format = "", na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_datetime(format), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_datetime
 #' @export
-parse_date <- function(x, format = "", na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_date(format), na = na, locale = locale)
+parse_date <- function(x, format = "", na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_date(format), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_datetime
 #' @export
-parse_time <- function(x, format = "", na = c("", "NA"), locale = default_locale()) {
-  parse_vector(x, col_time(format), na = na, locale = locale)
+parse_time <- function(x, format = "", na = c("", "NA"), locale = default_locale(), trim_ws = TRUE) {
+  parse_vector(x, col_time(format), na = na, locale = locale, trim_ws = trim_ws)
 }
 
 #' @rdname parse_datetime