+++ src/TokenizerWs.cpp
@@ -37,11 +37,8 @@ std::pair<double, size_t> TokenizerWs::progress() {
 }
 
 Token TokenizerWs::nextToken() {
-  if (cur_ == end_)
-    return Token(TOKEN_EOF, 0, 0);
-
-  // Check for comments only at start of line
-  while (cur_ != end_ && col_ == 0 && isComment(cur_)) {
+  // Check for comments and empty lines only at start of line
+  while (cur_ != end_ && col_ == 0 && (isComment(cur_) || isEmpty())) {
     // Skip rest of line
     while (cur_ != end_ && *cur_ != '\n' && *cur_ != '\r') {
       ++cur_;
@@ -52,6 +49,8 @@ Token TokenizerWs::nextToken() {
     }
     curLine_ = cur_;
   }
+  if (cur_ == end_)
+    return Token(TOKEN_EOF, 0, 0);
 
   // Find start of field
   SourceIterator fieldBegin = cur_;
@@ -92,4 +91,8 @@ bool TokenizerWs::isComment(const char* cur) const {
 
   boost::iterator_range<const char*> haystack(cur, end_);
   return boost::starts_with(haystack, comment_);
+}
+
+bool TokenizerWs::isEmpty() const {
+  return cur_ == end_ || *cur_ == '\r' || *cur_ == '\n';
 }