+++ src/TokenizerWs.h
@@ -29,6 +29,7 @@ private:
   Token fieldToken(SourceIterator begin, SourceIterator end, bool hasNull);
 
   bool isComment(const char* cur) const;
+  bool isEmpty() const;
 };
 
 #endif