+++ NEWS.md
@@ -1,10 +1,61 @@
-# readr 1.1.0.9000
+# readr 1.2.0
 
-* Added function `read_lines_raw_chunked()` (@gergness)
+This release was focused on bug fixes rather than new features.
 
+## Breaking Changes
+
+### Integer column guessing
+
+readr functions no longer guess columns are of type integer, instead these
+columns are guessed as numeric. Because R uses 32 bit integers and 64 bit
+doubles all integers can be stored in doubles, guaranteeing no loss of
+information. This change was made to remove errors when numeric columns were
+incorrectly guessed as integers. If you know a certain column is an integer and
+would like to read them as such you can do so by specifying the column type
+explicitly with the `col_types` argument.
+
+### Blank line skipping
+
+readr now always skips blank lines automatically when parsing, which may change
+the number of lines you need to pass to the `skip` parameter. For instance if
+your file had a one blank line then two more lines you want to skip previously
+you would pass `skip = 3`, now you only need to pass `skip = 2`.
+
+## New features
+
+* The `parse_*()` functions and `read_fwf()` gain a `trim_ws` argument to
+  control whether the fields should be trimmed before parsing (#636, #735).
+* `parse_number()` now parses numbers in scientific notation using `e` and `E`
+  (#684, @sambrady3).
+* Add `write_excel_csv2()` function to allow writing csv files with comma as a
+  decimal separator and semicolon as a column separator (#753, @olgamie).
+* `read_*()` files now support reading from the clipboard by using `clipboard()` (#656).
+* `write_file()` gains a `sep` argument, to specify the line separator (#665).
+* Allow files to be read via FTP over SSH by recognising `sftp` as a URL protocol (#707, @jdeboer).
+* `parse_date*() accepts `%a` for local day of week (#763, @tigertoes).
+* Added function `read_lines_raw_chunked()` (#710, @gergness)
+
+## Bug Fixes
+
+* `format_*()` now explicitly marks it's output encoding as UTF-8 (#697).
+* `read_delim()` now ignores whitespace between the delimiter and quoted fields (#668).
+* `read_table2()` now properly ignores blank lines at the end of a file like
+  `read_table()` and `read_delim()` (#657).
+* `read_delim()`, `read_table()` and `read_table()` now skip blank lines at the
+  start of a file (#680, #747).
+* `guess_parser()` now guesses a logical type for columns which are all
+  missing. This is useful when binding multiple files together where some files
+  have missing columns. (#662).
 * Column guessing will now never guess an integer type. This avoids issues
   where double columns are incorrectly guessed as integers if they have only
   integer values in the first 1000 (#645, #652).
+* `read_*()` now converts string `file`s to UTF-8 before parsing, which is convenient for non-UTF-8 platforms
+  in most cases (#730, @yutannihilation).
+* `write_csv()` writes integers up to 10^15 without scientific notation (#765, @zeehio)
+* `read_*()` no longer throws a "length of NULL cannot be changed" warning when
+  trying to resize a skipped column (#750, #833).
+* `read_*()` now handles non-ASCII paths properly with R >=3.5.0 on Windows (#838, @yutannihilation).
+
 # readr 1.1.1
 
 * Point release for test compatibility with tibble v1.3.1.